Although this is the first commit, it is a 2 week-ish learning progress.
Noticeable learning outcomes, I created a largestNumber.py file that takes integer input from the user and asks for more input. When there is no more input, it prints the smallest and largest numbers or equal if none.
It also prints the sorted set.

An improvement to the above file named largestNumberWhile.py uses the while loop and significantly reduces the lines of code and improves the experience for the user