Started learning about dictionaries. Dictionaries are Objects in javascript.
In largestNumberWhile.py, the challenge was to take a number from input, add it to a list and do things with it.
Learning dictionaries, I tried to figure a way to take input and add to the dictionary, but the challenge was dictionaries are a Key Value pair.
I could have requested two inputs but wanted a tougher challenge.

I wanted to have a dictionary that would tell me the number of times a name occurred.

When you run this program, it takes a name as input, stores it in a list. When you terminate the program, it prints out a dict with the names as the key and the number of occurence as the value