My previous push on the same file had a logic error
The task was to return a dictionary of the key with the largest value. But in error, I sorted the dicitonary by max and returned the value. This is wrong as dictionaries aren't lists so max methods only return the key comparison.

I went ahead to use the max method on the dictionary by querying it with .get for the key.
Then assigned the value of the key found.

Boom, now returns the word with the most appearances in the file.

Not sure why nor do I understand why it does, but it is working as expected. This time, I tested it across multiple files